Deep_Seek/nightowl2/requirements.txt
rich==13.7.0
requests==2.31.0
beautifulsoup4==4.12.2
psutil==5.9.5
pyyaml==6.0.1
jinja2==3.1.2
flask==3.0.0
boto3==1.34.0
botocore==1.34.0
stem==1.8.2
networkx==3.2.1
numpy==1.26.0
aiohttp==3.9.1
playwright==1.39.0
azure-identity==1.15.0
azure-mgmt-resource==23.1.1
google-cloud-resource-manager==1.10.4
sentence-transformers==2.5.1
scikit-learn==1.4.2
web3==6.15.1